+++ packages/origin-backend-app/src/mail/mail.module.ts
@@ -1,29 +1,17 @@
 import { HandlebarsAdapter, MailerModule } from '@nestjs-modules/mailer';
 import { Module } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
+import { MandrillTransport } from 'mandrill-nodemailer-transport';
 
 import { MailService } from './mail.service';
 
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const mandrillTransport = require('nodemailer-mandrill-transport');
 @Module({
     imports: [
         MailerModule.forRootAsync({
             useFactory: async (configService: ConfigService) => {
                 return {
-                    transport: mandrillTransport({
-                        auth: {
-                            apiKey: configService.get<string>('MANDRILL_API_KEY')
-                        },
-                        // eslint-disable-next-line camelcase
-                        from_email: configService.get<string>('EMAIL_FROM'),
-                        // eslint-disable-next-line camelcase
-                        from_name: configService.get<string>('EMAIL_FROM_NAME'),
-                        headers: {
-                            'Reply-To': configService.get<string>('EMAIL_REPLY_TO')
-                        },
-                        merge: true,
-                        tags: ['origin', 'no-reply']
+                    transport: new MandrillTransport({
+                        apiKey: configService.get<string>('MANDRILL_API_KEY')
                     }),
                     defaults: {
                         from: `"${configService.get<string>(